# BAAS-23655: Remove internal goja tick counter