https://jira.mongodb.org/browse/BAAS-23655

This rolls back this [commit](https://github.com/Calvinnix/goja/commit/f95411d186f2183429261010a860c604371ac6a7) from the maintainer that was fixing a bug where Runtime.interrupt was never called.

This change would fail `TestRunLoopPreempt` but we haven't ran that test since it was commented out [here](https://github.com/Calvinnix/goja/commit/c84a6965a8e80876c246809bd7e747004402e761)

I'm making an assumption that our custom interrupt logic resolves the need for calling runtime.Gosched periodically. If we see issues, worst case scenario we could add runtime.Gosched inside of waitOneTick.

current
```
goja (realm)$ go test -bench=^BenchmarkEmpty -benchtime 10000000x
goos: darwin
goarch: arm64
pkg: github.com/dop251/goja
BenchmarkEmptyLoop-10           10000000              5953 ns/op
PASS
ok      github.com/dop251/goja  60.116s
```

new
```
goja (BAAS-23655_removeInternalGojaTickCounter)$ go test -bench=^BenchmarkEmpty -benchtime 10000000x
goos: darwin
goarch: arm64
pkg: github.com/dop251/goja
BenchmarkEmptyLoop-10           10000000              5802 ns/op
PASS
ok      github.com/dop251/goja  58.620s
```

With this change we are seeing ~2.5% performance improvement specifically for the loop in vm.run(). This will get drowned out a bit depending on what customers are doing in their functions, but we will still probably see a nice improvement (see pprof investigation from ticket).